Making requested changes Tweaking fetching data in useEffects Moving constants to a single file Better error handling
