--- src/constants/apiNumOfItems.js
@@ -1,1 +1,0 @@
-export const NUM_OF_ITEMS = 325;
