+++ src/constants/constants.js
@@ -1,1 +1,3 @@
 export const BREWDOG_API = 'https://api.punkapi.com/v2/beers';
+
+export const NUM_OF_ITEMS = 325;
