+++ src/pages/beer-info/Index.jsx
@@ -2,7 +2,7 @@ import { useEffect, useState } from 'react';
 import { useParams } from 'react-router-dom';
 
 import fetchFromParams from '../../domain/fetchAPI/fetchFromParams';
-import { BREWDOG_API } from '../../constants/apiUrl';
+import { BREWDOG_API } from '../../constants/constants';
 import genericBottle from '../../assets/generic-bottle.png';
 
 export default function BeerInfo() {
@@ -10,48 +10,45 @@ export default function BeerInfo() {
   const { id } = useParams();
 
   useEffect(() => {
-    const fetchBeer = async () => {
-      const data = await fetchFromParams(BREWDOG_API, `/${id}`);
-      setBeer(data);
-    };
-
-    fetchBeer();
+    fetchFromParams(BREWDOG_API, `/${id}`).then((result) => {
+      setBeer(result[0]);
+    });
   }, [id]);
 
   return (
     <>
       {beer !== null && (
         <>
-          <h4>{beer[0].id}</h4>
-          <h3>{beer[0].name.toUpperCase()}</h3>
-          <p>{beer[0].tagline.toUpperCase()}</p>
+          <h4>{beer.id}</h4>
+          <h3>{beer.name.toUpperCase()}</h3>
+          <p>{beer.tagline.toUpperCase()}</p>
           <i>
-            <small>{beer[0].brewers_tips}</small>
+            <small>{beer.brewers_tips}</small>
           </i>
           <br />
           <br />
           <br />
-          {beer[0].image_url !== null ? (
-            <img src={beer[0].image_url} height="235px" width="auto" alt="beer image" />
+          {beer.image_url !== null ? (
+            <img src={beer.image_url} height="235px" width="auto" alt="beer image" />
           ) : (
             <img src={genericBottle} height="235px" width="auto" alt="beer image" />
           )}
-          <p>{beer[0].description}</p>
+          <p>{beer.description}</p>
           <section>
             <b>Hops: </b>
-            {beer[0].ingredients.hops.map((item, index) => (
+            {beer.ingredients.hops.map((item, index) => (
               <span key={index}>{item.name}, </span>
             ))}
           </section>
           <section>
             <b>Malt: </b>
-            {beer[0].ingredients.malt.map((item) => (
+            {beer.ingredients.malt.map((item) => (
               <span key={item.name}>{item.name}, </span>
             ))}
           </section>
           <section>
             <b>Yeast: </b>
-            <span>{beer[0].ingredients.yeast}</span>
+            <span>{beer.ingredients.yeast}</span>
           </section>
         </>
       )}
